Ajout de la barre de vie

+++ Assets/BarreDeVie.cs
@@ -13,7 +13,7 @@ public class BarreDeVie : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        int x = (gameObject.transform.parent.transform.parent.GetComponent<AttaqueMonstreRose>().nombreDeVie / gameObject.transform.parent.transform.parent.GetComponent<AttaqueMonstreRose>().nombreDeVieMax) * 5;
+        float x = (gameObject.transform.parent.transform.parent.GetComponent<AttaqueMonstreRose>().nombreDeVie / gameObject.transform.parent.transform.parent.GetComponent<AttaqueMonstreRose>().nombreDeVieMax) * 5f;
         gameObject.transform.localScale = new Vector3(x, 0.5f, 1f);
     }
 }
